requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
fiona==1.10.1
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
idna==3.10
networkx==3.3
numpy==1.26.4
osmnx==1.9.4
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whatsapp-api-client-python==0.0.47
whatsapp-chatbot-python==0.9.3
